Move contracts information to the new contract folder
